fix: split WineTests.swift to fix file_length swiftlint violation
Move WineLogRetentionTests to separate file to bring WineTests.swift
under the 400 line limit enforced by swiftlint --strict in CI.
